Remove canvas_extension and move drawDashedLine() to canvas_wrapper class
