+++ lib/src/utils/canvas_wrapper.dart
@@ -1,6 +1,7 @@
 import 'dart:ui';
 
 import 'package:fl_chart/fl_chart.dart';
+import 'package:fl_chart/src/extensions/path_extension.dart';
 import 'package:fl_chart/src/utils/utils.dart';
 import 'package:flutter/cupertino.dart' hide Image;
 
@@ -116,5 +117,15 @@ class CanvasWrapper {
     );
     drawCallback();
     restore();
+  }
+
+  /// Draws a dashed line from passed in offsets
+  void drawDashedLine(
+      Offset from, Offset to, Paint painter, List<int>? dashArray) {
+    var path = Path();
+    path.moveTo(from.dx, from.dy);
+    path.lineTo(to.dx, to.dy);
+    path = path.toDashedPath(dashArray);
+    drawPath(path, painter);
   }
 }
